+++ package.json
@@ -13,7 +13,8 @@
     "fastclick": "^1.0.6",
     "js-sha512": "^0.7.1",
     "muse-ui": "^2.1.0",
-    "nuxt": "^1.0.0-rc11"
+    "nuxt": "^1.0.0-rc11",
+    "vue-i18n": "^7.3.4"
   },
   "devDependencies": {
     "babel-eslint": "^8.1.2",
