+++ src/locales/en.json
@@ -1,0 +1,3 @@
+{
+
+}
