+++ src/locales/zh-CN.json
@@ -1,0 +1,3 @@
+{
+
+}
