+++ src/store/index.js
@@ -1,0 +1,8 @@
+import { Store } from 'vuex'
+import state from './state'
+import mutations from './mutations'
+
+export default () => new Store({
+  state,
+  mutations
+})
