+++ src/store/mutations.js
@@ -1,0 +1,7 @@
+export default {
+  SET_LANG (state, locale) {
+    if (state.locales.includes(locale)) {
+      state.locale = locale
+    }
+  }
+}
