+++ src/store/state.js
@@ -1,0 +1,4 @@
+export default {
+  locales: ['en', 'zh-CN'],
+  locale: 'en'
+}
